Add zero before hr/min if its less than 10
new Date().getHours(), new Date().getMinutes() would return the integer value which omits preceding zero. But, time input value is in hh:mm format, that means it will add a zero if the values are less than 10. As a result, the condition on line 40 would sometime be false.

To fix the issue,  getCurrentTIme() needs to add preceding zero if numbers are leass than 10.